requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
coverage==7.13.5
cryptography==46.0.7
distlib==0.4.0
dj-database-url==3.1.2
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.3.1
filelock==3.24.2
gunicorn==21.2.0
idna==3.11
iniconfig==2.3.0
packaging==26.0
platformdirs==4.9.2
pluggy==1.6.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.9
pycparser==3.0
Pygments==2.20.0
PyJWT==2.11.0
pytest==9.0.2
pytest-cov==7.1.0
pytest-django==4.12.0
python-decouple==3.8
pytz==2025.2
requests==2.31.0
sqlparse==0.5.5
tomli==2.4.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
waitress==3.0.2
whitenoise==6.6.0
Pillow==10.4.0
setuptools==68.0.0
